+++ benchmarks/flask_simple/requirements_scenario.txt
@@ -1,3 +1,1 @@
 flask==3.0.0
-gunicorn==20.1.0
-requests==2.31.0
